basic refactor complete - need to add imports module
