+++ cassie-hs/cassie-hs.cabal
@@ -50,7 +50,6 @@ library
                         Data.Cassie.Rules.Substitute,
                         Data.Cassie.Solver.Context,
                         Data.Cassie.Solver.EqSolve,
-                        Data.Cassie.Solver.Imports,
                         Data.Cassie.Solver.Internal,
                         Data.Cassie.Structures.Internal,
                         Data.Cassie.Structures.Instances.Complex,
