--- cassie-hs/src/Data/Cassie/Solver/Imports.hs
@@ -1,7 +1,0 @@
-{-# LANGUAGE Safe #-}
-module Data.Cassie.Solver.Imports
-    (
-    ) where
-
-import Data.Cassie.Parser.Lang 
-
